🚀Extracted dynamic form to seperate component, refactors and invoice-card text overflow fix🚀
